requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==2.1.1
requests==2.32.3
google-generativeai==0.8.1
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.13
python-dotenv==1.0.1
